collection_numbers: move CollectionNumberForm to Views/
Components::CollectionNumberForm -> Views::Controllers::CollectionNumbers::Form
Mechanical refactor per the views/components rule (#4324, #4331).
