+++ app/views/controllers/collection_numbers/edit.html.erb
@@ -8,7 +8,7 @@ add_context_nav(collection_number_form_edit_tabs(
 
 <div class="row">
   <div class="col-xs-12 col-sm-7">
-    <%= render(Components::CollectionNumberForm.new(
+    <%= render(Views::Controllers::CollectionNumbers::Form.new(
           @collection_number, back: @back
         )) %>
   </div>
